+++ e2e/test/scenarios/embedding/sdk-iframe-embedding-setup/helpers/index.ts
@@ -1,0 +1,45 @@
+import type { Dashboard, RecentItem } from "metabase-types/api";
+
+type RecentActivityIntercept = {
+  response: Cypress.Response<{ recents: RecentItem[] }>;
+};
+
+type DashboardIntercept = {
+  response: Cypress.Response<Dashboard>;
+};
+
+export const getPreviewIframe = () =>
+  cy
+    .get("iframe")
+    .should("be.visible")
+    .its("0.contentDocument")
+    .should("exist")
+    .its("body")
+    .should("not.be.empty");
+
+export const getEmbedSidebar = () => cy.findByRole("complementary");
+
+export const visitNewEmbedPage = () => {
+  cy.visit("/embed/new");
+  cy.wait("@dashboard");
+};
+
+export const assertRecentItemName = (
+  model: "dashboard" | "card",
+  resourceName: string,
+) => {
+  cy.get<RecentActivityIntercept>("@recentActivity").should((intercept) => {
+    const recentItem = intercept.response?.body.recents?.filter(
+      (recent) => recent.model === model,
+    )?.[0];
+
+    expect(recentItem.name).to.be.equal(resourceName);
+  });
+};
+
+export const assertDashboard = ({ id, name }: { id: number; name: string }) => {
+  cy.get<DashboardIntercept>("@dashboard").should((intercept) => {
+    expect(intercept.response?.body.id).to.be.equal(id);
+    expect(intercept.response?.body.name).to.be.equal(name);
+  });
+};
